+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 * VFS-9312 Removed $.css() usages
 * VFS-9313 Removed ":hidden" and ":visible" jQuery selectors usages
 * VFS-9310 Removed $.parents() usages
+* VFS-8656 Removed jQuery usage from login-related endpoints
 * VFS-10107 Fixed smooth scroll in Chrome
 * VFS-9795 Fixed user logout from Onezone GUI after remove from cluster members
 * VFS-9623 Increased speed of SASS compilation and fixed its unnecessary recompilation